User and Admin Routes and rate limiting

+++ controllers/jobsController.js
@@ -2,6 +2,7 @@ const Job = require('../models/jobs')
 const ErrorHandler = require('../utils/errorHandler')
 const catchAsyncErrors = require('../middlewares/catchAsyncErrors');
 const APIFilters = require('../utils/apiFilters');
+const path = require('path');
 
 //Get All jobs => /api/v1/jobs
 exports.getJobs = catchAsyncErrors(async (req, res, next) => {
@@ -26,17 +27,13 @@ exports.getJobs = catchAsyncErrors(async (req, res, next) => {
 exports.getJob = catchAsyncErrors(async (req, res, next) => {
 
     const job = await Job.find({ $and: [{ _id: req.params.id }, { slug: req.params.slug }] })
-    // .populate({
-    //     path: 'user',
-    //     select: 'name'
-    // });
+        .populate({
+            path: 'user',
+            select: 'name'
+        });
 
     if (!job || job.length === 0) {
-        // return next(new ErrorHandler('Job not found', 404));
-        return res.status(404).json({
-            success: false,
-            message: 'Job not found'
-        });
+        return next(new ErrorHandler('Job not found', 404));
     }
 
     res.status(200).json({
@@ -67,9 +64,9 @@ exports.updateJob = catchAsyncErrors(async (req, res, next) => {
     }
 
     // Check if the user is owner
-    // if (job.user.toString() !== req.user.id && req.user.role !== 'admin') {
-    //     return next(new ErrorHandler(`User(${req.user.id}) is not allowed to update this job.`))
-    // }
+    if (job.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorHandler(`User(${req.user.id}) is not allowed to update this job.`))
+    }
 
     job = await Job.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
@@ -92,19 +89,19 @@ exports.deleteJob = catchAsyncErrors(async (req, res, next) => {
     }
 
     // Check if the user is owner
-    // if (job.user.toString() !== req.user.id && req.user.role !== 'admin') {
-    //     return next(new ErrorHandler(`User(${req.user.id}) is not allowed to delete this job.`))
-    // }
+    if (job.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorHandler(`User(${req.user.id}) is not allowed to delete this job.`))
+    }
 
     // Deleting files associated with job
 
-    // for (let i = 0; i < job.applicantsApplied.length; i++) {
-    //     let filepath = `${__dirname}/public/uploads/${job.applicantsApplied[i].resume}`.replace('\\controllers', '');
+    for (let i = 0; i < job.applicantsApplied.length; i++) {
+        let filepath = `${__dirname}/public/uploads/${job.applicantsApplied[i].resume}`.replace('\\controllers', '');
 
-    //     fs.unlink(filepath, err => {
-    //         if (err) return console.log(err);
-    //     });
-    // }
+        fs.unlink(filepath, err => {
+            if (err) return console.log(err);
+        });
+    }
 
     job = await Job.findByIdAndDelete(req.params.id);
 
@@ -140,5 +137,74 @@ exports.jobStats = catchAsyncErrors(async (req, res, next) => {
     res.status(200).json({
         success: true,
         data: stats
+    });
+});
+
+// Apply to job using Resume  =>  /api/v1/job/:id/apply
+exports.applyJob = catchAsyncErrors(async (req, res, next) => {
+    let job = await Job.findById(req.params.id).select('+applicantsApplied');
+
+    if (!job) {
+        return next(new ErrorHandler('Job not found.', 404));
+    }
+
+    // Check that if job last date has been passed or not
+    if (job.lastDate < new Date(Date.now())) {
+        return next(new ErrorHandler('You can not apply to this job. Date is over.', 400));
+    }
+
+    // Check if user has applied before
+    for (let i = 0; i < job.applicantsApplied.length; i++) {
+        if (job.applicantsApplied[i].id === req.user.id) {
+            return next(new ErrorHandler('You have already applied for this job.', 400))
+        }
+    }
+
+    // Check the files
+    if (!req.files) {
+        return next(new ErrorHandler('Please upload file.', 400));
+    }
+
+    const file = req.files.file;
+
+    // Check file type
+    const supportedFiles = /.docx|.pdf/;
+    if (!supportedFiles.test(path.extname(file.name))) {
+        return next(new ErrorHandler('Please upload document file.', 400))
+    }
+
+    // Check doucument size
+    if (file.size > process.env.MAX_FILE_SIZE) {
+        return next(new ErrorHandler('Please upload file less than 2MB.', 400));
+    }
+
+    // Renaming resume
+    file.name = `${req.user.name.replace(' ', '_')}_${job._id}${path.parse(file.name).ext}`;
+
+    file.mv(`${process.env.UPLOAD_PATH}/${file.name}`, async err => {
+        if (err) {
+            console.log(err);
+            return next(new ErrorHandler('Resume upload failed.', 500));
+        }
+
+        await Job.findByIdAndUpdate(req.params.id, {
+            $push: {
+                applicantsApplied: {
+                    id: req.user.id,
+                    resume: file.name
+                }
+            }
+        }, {
+            new: true,
+            runValidators: true,
+            useFindAndModify: false
+        });
+
+        res.status(200).json({
+            success: true,
+            message: 'Applied to Job successfully.',
+            data: file.name
+        })
+
     });
 });
